+++ Changelog.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 - Add optional school holidays to Netherlands calendar, by @Flix6x (#556).
+- Add optional carnival to Netherlands calendar.
 
 ## v12.1.0 (2020-10-16)